+++ src/OpenTelemetry.OpAmp.Client/Internal/Messages/ConnectionSettingsMessage.cs
@@ -2,10 +2,11 @@
 // SPDX-License-Identifier: Apache-2.0
 
 using OpAmp.Proto.V1;
+using OpenTelemetry.OpAmp.Client.Messages;
 
 namespace OpenTelemetry.OpAmp.Client.Internal.Listeners.Messages;
 
-internal class ConnectionSettingsMessage : IOpAmpMessage
+internal class ConnectionSettingsMessage : OpAmpMessage
 {
     public ConnectionSettingsMessage(ConnectionSettingsOffers connectionSettingsOffers)
     {