+++ src/OpenTelemetry.OpAmp.Client/CHANGELOG.md
@@ -2,6 +2,19 @@
 
 ## Unreleased
 
+* Add setting to configure the factory used to create `HttpClient` instances
+  used for the OpAMP Plain HTTP transport.
+  ([#3589](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/3589))
+
+* Add support for subscribing and unsubscribing to messages from the OpAMP server.
+  ([#3593](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/3593))
+
+* Clean up directories and namespaces for public API.
+  ([#3612](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/3612))
+
+* Expose public `RemoteConfigMessage`.
+  ([#3614](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/3614))
+
 ## 0.1.0-alpha.3
 
 Released 2025-Nov-13